+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Support to custom Thrift protocol with `config.protocol = :binary`.
 - Fix ThriftServer::OperationFailed not found error.
 - Implement Master Process to manage and protect Child Process, keep it running/restart/stop, like Unicorn.
+- Add Sails.reload! to reload cache classes.
+- Sails console start with IRB class, and support `reload!`,`service` methods in console.
 
 ## 0.1.0 / 2014-12-9